+++ imgui-examples/examples/test_window_impl.rs
@@ -435,7 +435,7 @@ fn show_test_window(ui: &Ui, state: &mut State, opened: &mut bool) {
                 ));
                 ui.spacing();
 
-                Slider::new(im_str!("Wrap width"), -20.0 ..= 600.0)
+                Slider::new(im_str!("Wrap width")).range(-20.0 ..= 600.0)
                     .display_format(im_str!("%.0f"))
                     .build(ui, &mut state.wrap_width);
 
@@ -852,7 +852,9 @@ fn show_example_menu_file<'a>(ui: &Ui<'a>, state: &mut FileMenuState) {
                     ui.text(format!("Scrolling Text {}", i));
                 }
             });
-        Slider::new(im_str!("Value"), 0.0..=1.0).build(ui, &mut state.f);
+        Slider::new(im_str!("Value"))
+            .range(0.0..=1.0)
+            .build(ui, &mut state.f);
 
         ui.input_float(im_str!("Input"), &mut state.f)
             .step(0.1)
@@ -887,7 +889,9 @@ fn show_example_app_auto_resize(ui: &Ui, state: &mut AutoResizeState, opened: &m
 Note that you probably don't want to query the window size to
 output your content because that would create a feedback loop.",
             );
-            Slider::new(im_str!("Number of lines"), 1..=20).build(ui, &mut state.lines);
+            Slider::new(im_str!("Number of lines"))
+                .range(1..=20)
+                .build(ui, &mut state.lines);
             for i in 0..state.lines {
                 ui.text(format!("{:2$}This is line {}", "", i, i as usize * 4));
             }